[Tests] Add test for local_load to #mma layout.
This is a TTGIR test for the bug worked around in PR #3561.

While we're here, we also do a minor cleanup in test_core.py.

We also need to fix the regex for parsing an MLIR file to find the individual
arguments.  Previously it would include the function's closing paren in the
argument type.

GPC: local-load-test